refactor: move Ledger device error types to onboard/types.ts (WA-3243)
Types-only move, no behaviour change. `LedgerDeviceErrorReason` and
`LedgerDeviceErrorInfo` now live in `services/onboard/types.ts`, matching the
sibling `onboard/trezor/types.ts`.

`LEDGER_ERROR_SOURCE` stays in `ledger-errors.ts` — it is runtime state, and
keeping it there leaves `types.ts` a pure-types module like its sibling. The
interface declares the field as the literal `'ledger-device'`, and the const is
annotated `LedgerDeviceErrorInfo['source']` so the two cannot drift apart
without a compile error. That marker is what lets `getLedgerDeviceError` find
the payload after viem re-wraps the error, so silent drift would disable the
feature.
